requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pymysql
cryptography
python-dotenv
redis
passlib[bcrypt]
pyjwt
